Allow computer vision to operate on multiple images at the same time

+++ src/features/external_tools/external_tool.py
@@ -7,18 +7,18 @@ from features.images.image_size_utils import normalize_image_size_category
 @dataclass(frozen = True)
 class CostEstimate:
     """All costs are in credits."""
-    input_1m_tokens: int | None = None
-    output_1m_tokens: int | None = None
-    search_1m_tokens: int | None = None
-    input_image_1k: int | None = None
-    input_image_2k: int | None = None
-    input_image_4k: int | None = None
-    input_image_8k: int | None = None
-    input_image_12k: int | None = None
-    output_image_1k: int | None = None
-    output_image_2k: int | None = None
-    output_image_4k: int | None = None
-    api_call: int | None = None
+    input_1m_tokens: float | None = None
+    output_1m_tokens: float | None = None
+    search_1m_tokens: float | None = None
+    input_image_1k: float | None = None
+    input_image_2k: float | None = None
+    input_image_4k: float | None = None
+    input_image_8k: float | None = None
+    input_image_12k: float | None = None
+    output_image_1k: float | None = None
+    output_image_2k: float | None = None
+    output_image_4k: float | None = None
+    api_call: float | None = None
     second_of_runtime: float | None = None
 
     def get_minimum_for(
